Moved functions to separate file so they can be tested.
